casted project size to int

+++ lib/screens/projects/add_project_page.dart
@@ -200,7 +200,7 @@ class _AddProjectPageState extends State<AddProjectPage> {
           AdaptiveText("Project Size: "),
           ProjectSizeSlider(
             onChanged: (value) {
-              newTask["projectSize"] = value;
+              newTask["projectSize"] = value.toInt();
             },
           ),
         ],
